fix: componentize houseForm fix dailyPrice spelling
