+++ src/components/Forms/HouseForm/servicesOptions.ts
@@ -49,7 +49,7 @@ export const defaultNoValues = {
   photos: null,
   state: '',
   city: '',
-  daylyPrice: '',
+  dailyPrice: '',
   singleBed: '',
   doubleBed: '',
   services: [],
@@ -64,7 +64,7 @@ export const  defaultHouseFormValues = {
       } , 
   city: '',
   state: '',
-  daylyPrice: '',
+  dailyPrice: '',
   singleBed: '',
   doubleBed: '',
   services: [],
